requirements.txt
Flask
Gunicorn
requests
Pillow

